requirements.txt
Django
django-heroku
djangorestframework
gunicorn
psycopg2
